Move endpoint class to credentials folder
