+++ test/Intacct/Credentials/EndpointTest.php
@@ -2,9 +2,9 @@
 /**
  * Copyright 2017 Intacct Corporation.
  *
- *  Licensed under the Apache License, Version 2.0 (the "License"). You may not
- *  use this file except in compliance with the License. You may obtain a copy
- *  of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"). You may not
+ * use this file except in compliance with the License. You may obtain a copy
+ * of the License at
  *
  * http://www.apache.org/licenses/LICENSE-2.0
  *
@@ -12,10 +12,9 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
  * express or implied. See the License for the specific language governing
  * permissions and limitations under the License.
- *
  */
 
-namespace Intacct;
+namespace Intacct\Credentials;
 
 use InvalidArgumentException;
